Add files via upload
OVERVIEW OF THE CODE
__________________________________
PHYSICS BEHIND THE CODE:
The main purpose of the code is find the driving forces for the outflows from protostellar systems.
It is still not very understood how these outflows gets ejected from the protostellar systems. However, for sometime
now, the main idea became that ejections have been associated with magnetic fields. the current understanding is that
magnetic fields can extract angular momentum form the system and be able to launch these outflows. In this research,
we are trying to find the net forces on these outflows, to determine the driving forces on these outflows, and explain
their origin and ejections. The magnetic fields, velocity fields, Temperature, density and the rest of the parameters
were and published in the simulation that is mentioned in this paper: https://arxiv.org/abs/2401.04260.
The forces that are considered are the gravitational force, the thermal pressure gradient, and the Lorentz force
__________________________________
INPUT AND PROCESS:
This main code imports several files that should be in the same directory as this file. All of these files have
functions that complete each other are used in this code. All the files should be in Excel '.xlsx' format.
There are files for:
    - Calculating gravitational force
    - Calculating pressure gradient force
    - Calculating Lorentz force
    - Computing the Net force
    - Transform the data from Cartesian to Cylindrical coordinates
    - Taking the azimuthal averaged quantities
    - Plotting the data as vector fields
The simulation has a cube of grid cells 64x64x64. There are different set of resolution, which makes the grid cell
smaller, and it is like zooming into the center of the cube, where the protostar is.
the code uses the method of finite differences of the second order to calculate the gradient of the variables.
For plotting (for now), the code is using azimuthal angle averaged over the quantities, so we can plot this in the
r-z plane.
__________________________________
OUTPUT AND RESULTS:
The code outputs plots for each individual force and the net force in total, along with saving files in the same
directory of the code that carries all the data for the forces, in cartesian and cylindrical coordinates.
__________________________________
RUNING THE CODE:
To run the code, run the file 'main.py' and make sure that all the other files that are imported are in the same directory as the code. The code should take some time to give the results, since it is doing multiple calculations such as finite differences and plotting.